Enhance DEBUG_PRINT macro with stderr option
If DEBUG_PRINT_TO_STDERR is set to 1, the DEBUG_PRINT output will be sent to stderr. If not, then it will be sent to stdout.

Sending it to stdout for example-3 creates problems as PCM data is expected. Subsequently by default cmake sets DEBUG_PRINT_TO_STDERR to 1 to send DEBUG_PRINT to stderr